fix code coverage
