+++ src/test/java/pl/paweln/codility/sorting/algorithm/HeapSort2Test.java
@@ -4,10 +4,10 @@ import org.junit.Before;
 import pl.paweln.codility.core.SolutionFactory;
 import pl.paweln.codility.core.SolutionInputParams;
 
-public class QuickSortTest extends SortingAlgorithmBase {
+public class HeapSort2Test extends SortingAlgorithmBase {
     @Before
     public void setup() {
-        this.codilitySolution = this.solutionFactory.getSolution(SolutionFactory.CodilityTask.SORTING_QUICK_SORT);
+        this.codilitySolution = this.solutionFactory.getSolution(SolutionFactory.CodilityTask.SORTING_HEAP_SORT_2);
         this.paramsBuilder = new SolutionInputParams.SolutionInputParamsBuilder();
     }
 }
